requirements.txt
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.1.4
crispy-bootstrap5==2023.10
cryptography==42.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.7
Django==4.2.10
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-environ==0.11.2
django-humanize==0.1.2
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
EasyProcess==1.1
entrypoint2==1.1
executing==2.0.1
filelock==3.13.1
humanize==4.9.0
idna==3.6
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mss==9.0.1
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pbr==6.0.0
phonenumbers==8.13.27
Pillow==10.1.0
pipenv==2023.11.15
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.16.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
stack-data==0.6.3
stevedore==5.1.0
tornado==6.3.3
traitlets==5.13.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
wcwidth==0.2.9
